Multiple changes for printout, one fix
I tried to be as non-changing as possible and adding printouts
everywhere that matters.

I fixed 1 bug concerning the width of border tiles.

I fixed 1 potential bug cocerning sizes of buffers. This one is likely
to cause problems, right now. I apologize for too many changes in this
commit, try to be better in the future.